requirements.txt
requests
psycopg2-binary
python-dotenv
tenacity
pydantic
sqlalchemy
pydantic-settings